领取客户

+++ src/main/java/com/fjs/cronus/model/BusinessPool.java
@@ -28,7 +28,7 @@ public class BusinessPool {
     @ApiModelProperty(value = "媒体客户数量统计表id")
     private Integer mediaCustomerCountId;
     @ApiModelProperty(value = "核算方法(1-预购单价, 2-佣金比例, 3-放款比例, 4-预购单价&佣金比例, 5-预购单价&放款比例)")
-    private BigDecimal accountingMethod;
+    private Integer accountingMethod;
     @ApiModelProperty(value = "预购单价")
     private BigDecimal prepurchasePrice;
     @ApiModelProperty(value = "佣金比例")
@@ -108,11 +108,11 @@ public class BusinessPool {
         this.mediaCustomerCountId = mediaCustomerCountId;
     }
 
-    public BigDecimal getAccountingMethod() {
+    public Integer getAccountingMethod() {
         return accountingMethod;
     }
 
-    public void setAccountingMethod(BigDecimal accountingMethod) {
+    public void setAccountingMethod(Integer accountingMethod) {
         this.accountingMethod = accountingMethod;
     }
